Restyled Tables
Now the lobby table and the home lastGames Table use one table-style.css file. So the tables are allways the styled the same but the content can be different. This will also be applied to the Lobby Table later

Styling includes:
Green Header with white font
Table Row Styling with grey background color
Table Data Styling with the same padding as the header (not to bottom and right) and centered text in header and tabledata