Setins restyled
I restyled the settings to have a navbar on the left and the content on the right.
The next step is to lengthen the navbar from top to bottom. Also to center the content and only appear the navbar on the left. Also create background color navbar darker grey. content lighter grey. navbar green border.